interim checkin
+ Fixed enable of cold/multi-paritcipant merge issue
+ Fixed some new issues introduced related to types when updating flex UI

Spotted an issue with the transfer button that needs to be fixed. Previously when only cold transfer existed it was sufficient to disable the button the first time the transfer was clicked.

Even in the simple case of cold transfer there was additional complexity on if the transfer button should be available. It needed to keep track of all tasks that had been transferred and therefore shouldn't have the transfer button enabled. Because the task header is separate to the mesasge canvas it wasn't sufficient to have simple enable/disable local state that was set to disable after the button was clicked - instead local state needed to track task sids that had been transferred. This logic needs revisited now that multi-participant has been added